+++ deps/uvwasi/uvwasi.gyp
@@ -11,6 +11,7 @@
       'sources': [
         'src/clocks.c',
         'src/fd_table.c',
+        'src/path_resolver.c',
         'src/uv_mapping.c',
         'src/uvwasi.c',
         'src/wasi_rights.c',
